add validate to user's genre's id input

+++ retalApp2/routes/genres.js
@@ -14,7 +14,7 @@ router.get('/', async (req, res) => {
 });
 
 // only authonticated users can add new Genres
-router.post('/',  async (req, res) => {
+router.post('/', auth, async (req, res) => {
     const { error } = validate(req.body);
     if (error) return res.status(400).send(error.details[0].message);
 
@@ -42,7 +42,9 @@ router.put('/:id', async (req, res) => {
     res.send(genre);
 });
 
-router.delete('/:id',  async (req, res) => {
+router.delete('/:id', async (req, res) => {
+
+    if (!validateId(req.params.id)) return res.status(400).send("Invalid object id");
     const genre = await Genre.findByIdAndRemove(req.params.id, { useFindAndModify:false});
 
     if (!genre) return res.status(404).send('The genre with the given ID was not found.');
@@ -51,6 +53,7 @@ router.delete('/:id',  async (req, res) => {
 });
 
 router.get('/:id', async (req, res) => {
+    if (!validateId(req.params.id)) return res.status(400).send("Invalid object id");
     const genre = await Genre.findById(req.params.id);
 
     if (!genre) return res.status(404).send('The genre with the given ID was not found.');
